FORM-015: Add more tests, speed should not exceed top speed

+++ src/main/scala/com/formulaone/Race.scala
@@ -6,7 +6,7 @@ import com.formulaone.track.RaceTrack
 /**
  * Created by yash.datta on 29/03/16.
  */
-case class Car(topSpeed: Double, acc: Int, hf: Float) {
+case class Car(topSpeed: Double, acc: Int, hf: Double) {
   var speed: Double = 0
   var nitroUsage: Boolean = false
 
@@ -32,7 +32,7 @@ abstract class Race(val track: RaceTrack, val numTeams: Int)
   def init(): Unit = {
     var i = 1
     while (i <= numTeams) {
-      cars(i-1) = Car( ((150 + 10*i)*0.2777).toInt, 2*i, 0.8f)
+      cars(i-1) = Car( ((150 + 10*i)*0.2777), 2*i, 0.8)
       i += 1
     }
   }
